# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Core Api",
        "version": "1.0.0"
    },
    "basePath": "/Week13_Assignment/public/api",
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get categories list",
                "description": "Get categories list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "category": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create category",
                "description": "Create category",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "store successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "category": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing field"
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update Category",
                "description": "Update Category",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "formData",
                        "description": "category",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "category": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    },
                    "302": {
                        "description": "wrong validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete one specific category",
                "description": "Delete one specific category",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    },
                    "404": {
                        "description": "Category Not Found"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Posts list",
                "description": "Posts list",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "category": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "author": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create one post",
                "description": "Create one post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "category_id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "category": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    },
                    "302": {
                        "description": "wrong validation"
                    }
                }
            }
        },
        "/posts/{id}": {
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update one Post",
                "description": "Update one Post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "category_id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "category": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    },
                    "302": {
                        "description": "wrong validation"
                    }
                }
            }
        },
        "/post/{id}": {
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete one post",
                "description": "Delete one  post",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "interal server error (error inserting to database)"
                    },
                    "404": {
                        "description": "Post Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "Category": {
            "required": [
                "label",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "created at": {
                    "description": "created at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated at": {
                    "description": "updated at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}